# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TwilightCup Backend",
    "description": "黄昏杯（Twilight Cup）《人类一败涂地》1v1 速通比赛中转服务端。REST 接口用于账号/比赛管理与日志查询；选手·裁判·导播三端的实时通信走 WebSocket（协议见 docs/ws-protocol.md）。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "登录签发令牌",
        "description": "用用户名+口令换取 JWT 令牌，REST 与 WebSocket 均以该令牌鉴权。",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "账号列表",
        "description": "返回全部账号（不含口令哈希）。",
        "operationId": "list_admin_accounts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountOut"
                  },
                  "type": "array",
                  "title": "Response List Admin Accounts Get"
                }
              }
            }
          },
          "401": {
            "description": "未携带有效令牌"
          },
          "403": {
            "description": "需要管理员权限"
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "创建账号",
        "description": "管理员预创建选手/裁判/导播/管理员账号并分发。",
        "operationId": "create_admin_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "401": {
            "description": "未携带有效令牌"
          },
          "403": {
            "description": "需要管理员权限"
          },
          "409": {
            "description": "用户名已存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/matches": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "比赛列表",
        "description": "返回全部比赛。",
        "operationId": "list_admin_matches_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatchOut"
                  },
                  "type": "array",
                  "title": "Response List Admin Matches Get"
                }
              }
            }
          },
          "401": {
            "description": "未携带有效令牌"
          },
          "403": {
            "description": "需要管理员权限"
          }
        }
      },
      "post": {
        "tags": [
          "matches"
        ],
        "summary": "创建比赛",
        "description": "管理员创建一场比赛：指定赛制(BO)、单关计分方式、完整图池、指派双方选手/裁判/导播账号、开始倒计时延迟秒数。角色账号以用户名指定，服务端校验账号类型并解析为 id。win_threshold 省略时按 (bo//2)+1 推导。",
        "operationId": "create_admin_matches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOut"
                }
              }
            }
          },
          "401": {
            "description": "未携带有效令牌"
          },
          "403": {
            "description": "需要管理员权限"
          },
          "400": {
            "description": "角色账号不存在/类型不符/双方选手相同/分数非法"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCreate": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "登录用户名（唯一）"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "口令明文"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "展示名，如“选手A”"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/AccountType"
            },
            "type": "array",
            "title": "Roles",
            "description": "角色集合（可多选）：1=选手 2=裁判 3=导播 4=管理员"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password",
          "display_name",
          "roles"
        ],
        "title": "AccountCreate"
      },
      "AccountOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/AccountType"
            },
            "type": "array",
            "title": "Roles"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "roles",
          "display_name",
          "created_at"
        ],
        "title": "AccountOut"
      },
      "AccountType": {
        "type": "integer",
        "enum": [
          1,
          2,
          3,
          4
        ],
        "title": "AccountType",
        "description": "账号类型。"
      },
...
      "Category": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "picks": {
            "items": {
              "$ref": "#/components/schemas/Pick"
            },
            "type": "array",
            "title": "Picks"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Category",
        "description": "图池类别（仅展示分组，无程序逻辑）。"
      },
...
      "CollectionConfig": {
        "properties": {
          "raw": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "raw"
        ],
        "title": "CollectionConfig",
        "description": "关卡合集配置（插件定义格式，服务端原样存储，下发前把关卡 id 展开为名字）。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "登录用户名"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "口令明文（服务端仅存哈希）"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "Mappool": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/Category"
            },
            "type": "array",
            "title": "Categories"
          }
        },
        "type": "object",
        "title": "Mappool",
        "description": "图池：一场比赛所有可选项目的集合。"
      },
...
      "MatchCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "比赛名称"
          },
          "bo_format": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Bo Format",
            "description": "赛制 BO 数，如 9 表示 BO9"
          },
          "win_threshold": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Win Threshold",
            "description": "取胜所需分数；省略时按 (bo//2)+1 推导（BO9→5）"
          },
          "scoring_method": {
            "$ref": "#/components/schemas/ScoringMethod",
            "description": "单关项目成绩计算方式：1=最快 2=平均（仅对单关项目生效）"
          },
          "start_countdown_delay": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Countdown Delay",
            "description": "回合开始前的倒数秒数，默认 5",
            "default": 5
          },
          "ban_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Ban Count",
            "description": "每方选图 ban 数，默认 1",
            "default": 1
          },
          "protect_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Protect Count",
            "description": "每方选图 protect 数（0=无 protect，如深度赛），默认 1",
            "default": 1
          },
          "ct_tag_count": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Ct Tag Count",
            "description": "CT 词条数上限（0=本场禁用词条），默认 2",
            "default": 2
          },
          "mappool": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Mappool"
              },
              {
                "type": "null"
              }
            ],
            "description": "内联完整图池（与 mappool_id 二选一）"
          },
          "mappool_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mappool Id",
            "description": "引用图池库中的图池 id（优先于内联 mappool）"
          },
          "player_a": {
            "type": "string",
            "title": "Player A",
            "description": "选手 A 用户名"
          },
          "player_b": {
            "type": "string",
            "title": "Player B",
            "description": "选手 B 用户名"
          },
          "referee": {
            "type": "string",
            "title": "Referee",
            "description": "裁判用户名"
          },
          "director": {
            "type": "string",
            "title": "Director",
            "description": "导播用户名"
          }
        },
        "type": "object",
        "required": [
          "name",
          "bo_format",
          "scoring_method",
          "player_a",
          "player_b",
          "referee",
          "director"
        ],
        "title": "MatchCreate"
      },
...
      "MatchOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "bo_format": {
            "type": "integer",
            "title": "Bo Format"
          },
          "win_threshold": {
            "type": "integer",
            "title": "Win Threshold"
          },
          "scoring_method": {
            "$ref": "#/components/schemas/ScoringMethod"
          },
          "start_countdown_delay": {
            "type": "integer",
            "title": "Start Countdown Delay"
          },
          "ban_count": {
            "type": "integer",
            "title": "Ban Count"
          },
          "protect_count": {
            "type": "integer",
            "title": "Protect Count"
          },
          "ct_tag_count": {
            "type": "integer",
            "title": "Ct Tag Count"
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus"
          },
          "mappool": {
            "$ref": "#/components/schemas/Mappool"
          },
          "player_a_id": {
            "type": "string",
            "title": "Player A Id"
          },
          "player_b_id": {
            "type": "string",
            "title": "Player B Id"
          },
          "player_a_username": {
            "type": "string",
            "title": "Player A Username",
            "default": ""
          },
          "player_b_username": {
            "type": "string",
            "title": "Player B Username",
            "default": ""
          },
          "referee_id": {
            "type": "string",
            "title": "Referee Id"
          },
          "director_id": {
            "type": "string",
            "title": "Director Id"
          },
          "winner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner"
          },
          "tournament_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tournament Id"
          },
          "fixture_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixture Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "bo_format",
          "win_threshold",
          "scoring_method",
          "start_countdown_delay",
          "ban_count",
          "protect_count",
          "ct_tag_count",
          "status",
          "mappool",
          "player_a_id",
          "player_b_id",
          "referee_id",
          "director_id",
          "winner",
          "tournament_id",
          "fixture_id",
          "created_at",
          "started_at",
          "ended_at"
        ],
        "title": "MatchOut"
      },
      "MatchStatus": {
        "type": "integer",
        "enum": [
          0,
          1,
          2,
          3
        ],
        "title": "MatchStatus",
        "description": "比赛比赛状态（持久化层）。"
      },
...
      "Pick": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/PickType"
          },
          "retry_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Count"
          },
          "collection": {
            "$ref": "#/components/schemas/CollectionConfig"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "logo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "type",
          "collection"
        ],
        "title": "Pick",
        "description": "选图（= 一个项目 / 一个回合）。"
      },
      "PickType": {
        "type": "integer",
        "enum": [
          1,
          2
        ],
        "title": "PickType",
        "description": "选图项目类型。"
      },
...
      "ScoringMethod": {
        "type": "integer",
        "enum": [
          1,
          2
        ],
        "title": "ScoringMethod",
        "description": "单关项目成绩计算方式。"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT 令牌，用于 REST Bearer 与 WebSocket /ws/{token}"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "登录名（唯一）"
          },
          "account_type": {
            "$ref": "#/components/schemas/AccountType",
            "description": "主角色（最高优先级）"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/AccountType"
            },
            "type": "array",
            "title": "Roles",
            "description": "角色集合"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "account_id",
          "username",
          "account_type",
          "roles",
          "display_name"
        ],
        "title": "TokenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}